Refactor benchmarks to make room for new one
