+++ shared/src/main/scala-2.12/de/bbisping/eqfiddle/tool/Benchmark.scala
@@ -1,70 +1,13 @@
 package de.bbisping.eqfiddle.tool
 
-import de.bbisping.eqfiddle.spectroscopy.EdgeSpectroscopy
 import de.bbisping.eqfiddle.algo.AlgorithmLogging
 import de.bbisping.eqfiddle.algo.AlgorithmLogging.{LogRelation, LogRichRelation}
-import de.bbisping.eqfiddle.ccs.CCSSamples
-import de.bbisping.eqfiddle.ccs.Parser
-import de.bbisping.eqfiddle.ccs.Interpreter
-import de.bbisping.eqfiddle.ccs.Syntax
-import de.bbisping.eqfiddle.util.Interpreting
-import de.bbisping.eqfiddle.tool.model.NodeID
-import de.bbisping.eqfiddle.ts.WeakTransitionSystem
+import de.bbisping.eqfiddle.tool.benchmark.LTBTSDistinctions
 
-
-/**
- * Creates the distinguishing formulas that are used in the paper to illustrate the output of the algorithm.
- */
 object Benchmark extends App {
 
   AlgorithmLogging.loggingActive = true
   AlgorithmLogging.debugLogActive = false
 
-  val examplePairs = List(
-    ("L13", "R13"),
-    ("L16", "R16"),
-    ("L21", "R21"),
-    ("R24", "L24"),
-    ("L27", "R27"),
-    ("R31", "L31"),
-    ("L34", "R31"),
-    ("L38", "R24"),
-    ("L42", "R42"),
-    ("L50", "R50")
-  )
-
-  def arrowLabeling(o: Option[Syntax.Label]) = {
-    Interpreting.fromOption(o.map(_.name) orElse(Some("")))
-  }
-  def nodeLabeling(o: Option[Syntax.NodeDeclaration]) = {
-    Interpreting.fromOption(o.map(_.name) orElse(Some("")))
-  }
-  def actionIsOutput(a: String) = a.endsWith("!")
-  def actionToInput(a: String): String = if (actionIsOutput(a)) actionToInput(a.dropRight(1)) else a
-
-  val parser: Parser = new Parser(CCSSamples.ltbts1)
-  val parser.ParseSuccess(esDef, _) = parser.parse
-  val interpreter = new Interpreter(esDef, NodeID(_), arrowLabeling, nodeLabeling, actionToInput, actionIsOutput)
-  val Interpreting.Success(is) = interpreter.result(new WeakTransitionSystem(_, _, Set()))
-  val ltbtsSystem = is.asInstanceOf[WeakTransitionSystem[NodeID, String, String]]
-  
-  for ((n1s, n2s) <- examplePairs) {
-    val n1 = NodeID(n1s)
-    val n2 = NodeID(n2s)
-
-    println("--------------------")
-    esDef.getDeclaration(n1s) map (_.process) foreach (println(_))
-    esDef.getDeclaration(n2s) map (_.process) foreach (println(_))
-
-    val algo = new EdgeSpectroscopy(ltbtsSystem, List(n1, n2))
-    val result = algo.compute()
-
-    for {
-      res <- result.relationItems
-      if res.left == n1 && res.right == n2
-      (formula, obsClass, notions) <- res.distinctions
-    } {
-      println(formula.toString() + " ∈ " + notions.map(_.name).mkString(" ∩ "))
-    }
-  }
+  new LTBTSDistinctions(0).run()
 }
